add generate excelEquipmentOperationTimeReport

+++ src/Belaz.WeldingApp.RegistarApi.Domain/Entities/WeldingEquipmentInfo/WeldingEquipmentConditionTime.cs
@@ -15,7 +15,7 @@ public class WeldingEquipmentConditionTime : Entity
     /// <summary>
     /// Сколько оборудование находиться в данном состоянии в минутах
     /// </summary>
-    public int Time { get; set; } = 0;
+    public double Time { get; set; } = 0;
 
     public Guid WeldingEquipmentId { get; set; }
